level system

+++ src/Model/Enemy.java
@@ -2,7 +2,7 @@ package Model;
 
 public class Enemy extends Entity
 {
-	public Enemy(String nameIn, int hpIn, int dmgIn, int strIn, int dexIn, int conIn, int xpIn, int levelIn) {
+	public Enemy(String nameIn, int hpIn, int dmgIn, int strIn, int dexIn, int conIn, double xpIn, int levelIn) {
 		super(nameIn, hpIn, dmgIn, strIn, dexIn, conIn, xpIn, levelIn);
 		
 	}
